Adding submodule for comic-shanns source font, with a hacky twist
comic-mono-font, this repository's parent repo, is itself a fork of
comic-shanns, so github wouldn't let me fork it - but we need to track
the OG comic-shanns repo somehow.

This was accomplished by having a branch in this repo tracking the OG
comic-shanns repo's state, created by checking out the commit from when
comic-mono-font originally forked comic-shanns, then fast-forwading with
the OG comic-shanns as a remote.

so now, the comic-shanns branch on this repository is tracking exactly
the OG comic-shanns commits and state.

Topping it all of, we add this repo as a submodule to itself, tracking
that specific branch.